video fixed

+++ controllers/videoController.js
@@ -1,18 +1,20 @@
 import routes from "../routes";
 import Video from "../models/Video";
 
-//Home
+// Home
+
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({}).sort({ _id: -1 });
     res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
     console.log(error);
-    res.render("home", { pageTitle: "Home", videos });
+    res.render("home", { pageTitle: "Home", videos: [] });
   }
 };
 
-//Search
+// Search
+
 export const search = async (req, res) => {
   const {
     query: { term: searchingBy }
@@ -20,10 +22,7 @@ export const search = async (req, res) => {
   let videos = [];
   try {
     videos = await Video.find({
-      $or: [
-        { title: { $regex: searchingBy, $options: "i" } },
-        { description: { $regex: searchingBy, $options: "i" } }
-      ]
+      title: { $regex: searchingBy, $options: "i" }
     });
   } catch (error) {
     console.log(error);
@@ -31,7 +30,8 @@ export const search = async (req, res) => {
   res.render("search", { pageTitle: "Search", searchingBy, videos });
 };
 
-//Upload
+// Upload
+
 export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
 
@@ -40,27 +40,36 @@ export const postUpload = async (req, res) => {
     body: { title, description },
     file: { path }
   } = req;
-
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  console.log(`From postUpload Controller`);
+  console.log(`req.user: ${req.user}`);
+  console.log(`req.user.id: ${req.user.id}`);
+  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
-//Video Detail
+// Video Detail
+
 export const videoDetail = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
   }
 };
+
+// Edit Video
 
 export const getEditVideo = async (req, res) => {
   const {
@@ -68,7 +77,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -87,12 +100,19 @@ export const postEditVideo = async (req, res) => {
   }
 };
 
+// Delete Video
+
 export const deleteVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
